feat(pluginutils): add `exactRegex` and `prefixRegex`
